Session factory extender
