+++ maven-core/src/main/java/org/apache/maven/internal/aether/DefaultRepositorySystemSessionFactory.java
@@ -24,7 +24,6 @@ import javax.inject.Named;
 import java.io.File;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Collections;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.LinkedHashMap;
@@ -37,9 +36,6 @@ import org.apache.maven.RepositoryUtils;
 import org.apache.maven.api.services.TypeRegistry;
 import org.apache.maven.api.xml.XmlNode;
 import org.apache.maven.artifact.handler.manager.ArtifactHandlerManager;
-import org.apache.maven.artifact.repository.ArtifactRepository;
-import org.apache.maven.artifact.repository.Authentication;
-import org.apache.maven.bridge.MavenRepositorySystem;
 import org.apache.maven.eventspy.internal.EventSpyDispatcher;
 import org.apache.maven.execution.MavenExecutionRequest;
 import org.apache.maven.internal.xml.XmlNodeImpl;
@@ -47,6 +43,7 @@ import org.apache.maven.internal.xml.XmlPlexusConfiguration;
 import org.apache.maven.model.ModelBase;
 import org.apache.maven.repository.internal.MavenRepositorySystemUtils;
 import org.apache.maven.rtinfo.RuntimeInformation;
+import org.apache.maven.session.RepositorySystemSessionExtender;
 import org.apache.maven.settings.Mirror;
 import org.apache.maven.settings.Proxy;
 import org.apache.maven.settings.Server;
@@ -63,10 +60,6 @@ import org.eclipse.aether.RepositorySystemSession.SessionBuilder;
 import org.eclipse.aether.artifact.Artifact;
 import org.eclipse.aether.artifact.DefaultArtifact;
 import org.eclipse.aether.collection.VersionFilter;
-import org.eclipse.aether.repository.AuthenticationContext;
-import org.eclipse.aether.repository.AuthenticationSelector;
-import org.eclipse.aether.repository.ProxySelector;
-import org.eclipse.aether.repository.RemoteRepository;
 import org.eclipse.aether.repository.RepositoryPolicy;
 import org.eclipse.aether.repository.WorkspaceReader;
 import org.eclipse.aether.resolution.ResolutionErrorPolicy;
@@ -192,6 +185,8 @@ public class DefaultRepositorySystemSessionFactory {
 
     private final VersionScheme versionScheme;
 
+    private final Map<String, RepositorySystemSessionExtender> extenders;
+
     @SuppressWarnings("checkstyle:ParameterNumber")
     @Inject
     public DefaultRepositorySystemSessionFactory(
@@ -202,7 +197,8 @@ public class DefaultRepositorySystemSessionFactory {
             EventSpyDispatcher eventSpyDispatcher,
             RuntimeInformation runtimeInformation,
             TypeRegistry typeRegistry,
-            VersionScheme versionScheme) {
+            VersionScheme versionScheme,
+            Map<String, RepositorySystemSessionExtender> extenders) {
         this.artifactHandlerManager = artifactHandlerManager;
         this.repoSystem = repoSystem;
         this.workspaceRepository = workspaceRepository;
@@ -211,6 +207,7 @@ public class DefaultRepositorySystemSessionFactory {
         this.runtimeInformation = runtimeInformation;
         this.typeRegistry = typeRegistry;
         this.versionScheme = versionScheme;
+        this.extenders = extenders;
     }
 
     @Deprecated
@@ -220,9 +217,9 @@ public class DefaultRepositorySystemSessionFactory {
 
     @SuppressWarnings("checkstyle:methodLength")
     public SessionBuilder newRepositorySessionBuilder(MavenExecutionRequest request) {
-        SessionBuilder session = MavenRepositorySystemUtils.newSession(
+        SessionBuilder sessionBuilder = MavenRepositorySystemUtils.newSession(
                 repoSystem.createSessionBuilder(), new TypeRegistryAdapter(typeRegistry));
-        session.setCache(request.getRepositoryCache());
+        sessionBuilder.setCache(request.getRepositoryCache());
 
         Map<Object, Object> configProps = new LinkedHashMap<>();
         configProps.put(ConfigurationProperties.USER_AGENT, getUserAgent());
@@ -234,9 +231,9 @@ public class DefaultRepositorySystemSessionFactory {
         configProps.putAll(request.getSystemProperties());
         configProps.putAll(request.getUserProperties());
 
-        session.setOffline(request.isOffline());
-        session.setChecksumPolicy(request.getGlobalChecksumPolicy());
-        session.setUpdatePolicy(
+        sessionBuilder.setOffline(request.isOffline());
+        sessionBuilder.setChecksumPolicy(request.getGlobalChecksumPolicy());
+        sessionBuilder.setUpdatePolicy(
                 request.isNoSnapshotUpdates()
                         ? RepositoryPolicy.UPDATE_POLICY_NEVER
                         : request.isUpdateSnapshots() ? RepositoryPolicy.UPDATE_POLICY_ALWAYS : null);
@@ -248,20 +245,20 @@ public class DefaultRepositorySystemSessionFactory {
         errorPolicy |= request.isCacheTransferError()
                 ? ResolutionErrorPolicy.CACHE_TRANSFER_ERROR
                 : ResolutionErrorPolicy.CACHE_DISABLED;
-        session.setResolutionErrorPolicy(
+        sessionBuilder.setResolutionErrorPolicy(
                 new SimpleResolutionErrorPolicy(errorPolicy, errorPolicy | ResolutionErrorPolicy.CACHE_NOT_FOUND));
 
-        session.setArtifactDescriptorPolicy(new SimpleArtifactDescriptorPolicy(
+        sessionBuilder.setArtifactDescriptorPolicy(new SimpleArtifactDescriptorPolicy(
                 request.isIgnoreMissingArtifactDescriptor(), request.isIgnoreInvalidArtifactDescriptor()));
 
         VersionFilter versionFilter = buildVersionFilter((String) configProps.get(MAVEN_VERSION_FILTERS));
         if (versionFilter != null) {
-            session.setVersionFilter(versionFilter);
+            sessionBuilder.setVersionFilter(versionFilter);
         }
 
-        session.setArtifactTypeRegistry(RepositoryUtils.newArtifactTypeRegistry(artifactHandlerManager));
+        sessionBuilder.setArtifactTypeRegistry(RepositoryUtils.newArtifactTypeRegistry(artifactHandlerManager));
 
-        session.setWorkspaceReader(
+        sessionBuilder.setWorkspaceReader(
                 request.getWorkspaceReader() != null ? request.getWorkspaceReader() : workspaceRepository);
 
         DefaultSettingsDecryptionRequest decrypt = new DefaultSettingsDecryptionRequest();
@@ -286,7 +283,7 @@ public class DefaultRepositorySystemSessionFactory {
                     mirror.getMirrorOf(),
                     mirror.getMirrorOfLayouts());
         }
-        session.setMirrorSelector(mirrorSelector);
+        sessionBuilder.setMirrorSelector(mirrorSelector);
 
         DefaultProxySelector proxySelector = new DefaultProxySelector();
         for (Proxy proxy : decrypted.getProxies()) {
@@ -297,7 +294,7 @@ public class DefaultRepositorySystemSessionFactory {
                             proxy.getProtocol(), proxy.getHost(), proxy.getPort(), authBuilder.build()),
                     proxy.getNonProxyHosts());
         }
-        session.setProxySelector(proxySelector);
+        sessionBuilder.setProxySelector(proxySelector);
 
         // Note: we do NOT use WagonTransportConfigurationKeys here as Maven Core does NOT depend on Wagon Transport
         // and this is okay and "good thing".
@@ -390,7 +387,7 @@ public class DefaultRepositorySystemSessionFactory {
             configProps.put("aether.transport.wagon.perms.fileMode." + server.getId(), server.getFilePermissions());
             configProps.put("aether.transport.wagon.perms.dirMode." + server.getId(), server.getDirectoryPermissions());
         }
-        session.setAuthenticationSelector(authSelector);
+        sessionBuilder.setAuthenticationSelector(authSelector);
 
         Object transport = configProps.getOrDefault(MAVEN_RESOLVER_TRANSPORT_KEY, MAVEN_RESOLVER_TRANSPORT_DEFAULT);
         if (MAVEN_RESOLVER_TRANSPORT_DEFAULT.equals(transport)) {
@@ -420,12 +417,12 @@ public class DefaultRepositorySystemSessionFactory {
                     + MAVEN_RESOLVER_TRANSPORT_AUTO);
         }
 
-        session.setUserProperties(request.getUserProperties());
-        session.setSystemProperties(request.getSystemProperties());
-        session.setConfigProperties(configProps);
-        session.setIgnoreArtifactDescriptorRepositories(request.isIgnoreTransitiveRepositories());
+        sessionBuilder.setUserProperties(request.getUserProperties());
+        sessionBuilder.setSystemProperties(request.getSystemProperties());
+        sessionBuilder.setConfigProperties(configProps);
+        sessionBuilder.setIgnoreArtifactDescriptorRepositories(request.isIgnoreTransitiveRepositories());
 
-        session.setTransferListener(request.getTransferListener());
+        sessionBuilder.setTransferListener(request.getTransferListener());
 
         RepositoryListener repositoryListener = eventSpyDispatcher.chainListener(new LoggingRepositoryListener(logger));
 
@@ -434,19 +431,11 @@ public class DefaultRepositorySystemSessionFactory {
         if (recordReverseTree) {
             repositoryListener = new ChainedRepositoryListener(repositoryListener, new ReverseTreeRepositoryListener());
         }
-        session.setRepositoryListener(repositoryListener);
-
-        injectMirror(request.getRemoteRepositories(), request.getMirrors());
-        injectProxy(proxySelector, request.getRemoteRepositories());
-        injectAuthentication(authSelector, request.getRemoteRepositories());
-
-        injectMirror(request.getPluginArtifactRepositories(), request.getMirrors());
-        injectProxy(proxySelector, request.getPluginArtifactRepositories());
-        injectAuthentication(authSelector, request.getPluginArtifactRepositories());
+        sessionBuilder.setRepositoryListener(repositoryListener);
 
         String resolverDependencyManagerTransitivity = (String)
                 configProps.getOrDefault(MAVEN_RESOLVER_DEPENDENCY_MANAGER_TRANSITIVITY_KEY, Boolean.TRUE.toString());
-        session.setDependencyManager(
+        sessionBuilder.setDependencyManager(
                 new ClassicDependencyManager(Boolean.parseBoolean(resolverDependencyManagerTransitivity)));
 
         ArrayList<File> paths = new ArrayList<>();
@@ -458,9 +447,13 @@ public class DefaultRepositorySystemSessionFactory {
                     .map(File::new)
                     .forEach(paths::add);
         }
-        session.withLocalRepositoryBaseDirectories(paths);
+        sessionBuilder.withLocalRepositoryBaseDirectories(paths);
 
-        return session;
+        for (RepositorySystemSessionExtender extender : extenders.values()) {
+            extender.extend(request, sessionBuilder, mirrorSelector, proxySelector, authSelector);
+        }
+
+        return sessionBuilder;
     }
 
     private VersionFilter buildVersionFilter(String filterExpression) {
@@ -546,96 +539,5 @@ public class DefaultRepositorySystemSessionFactory {
         version = version.isEmpty() ? version : "/" + version;
         return "Apache-Maven" + version + " (Java " + System.getProperty("java.version") + "; "
                 + System.getProperty("os.name") + " " + System.getProperty("os.version") + ")";
-    }
-
-    private void injectMirror(List<ArtifactRepository> repositories, List<Mirror> mirrors) {
-        if (repositories != null && mirrors != null) {
-            for (ArtifactRepository repository : repositories) {
-                Mirror mirror = MavenRepositorySystem.getMirror(repository, mirrors);
-                injectMirror(repository, mirror);
-            }
-        }
-    }
-
-    private void injectMirror(ArtifactRepository repository, Mirror mirror) {
-        if (mirror != null) {
-            ArtifactRepository original = MavenRepositorySystem.createArtifactRepository(
-                    repository.getId(),
-                    repository.getUrl(),
-                    repository.getLayout(),
-                    repository.getSnapshots(),
-                    repository.getReleases());
-
-            repository.setMirroredRepositories(Collections.singletonList(original));
-
-            repository.setId(mirror.getId());
-            repository.setUrl(mirror.getUrl());
-
-            if (mirror.getLayout() != null && !mirror.getLayout().isEmpty()) {
-                repository.setLayout(original.getLayout());
-            }
-
-            repository.setBlocked(mirror.isBlocked());
-        }
-    }
-
-    private void injectProxy(ProxySelector selector, List<ArtifactRepository> repositories) {
-        if (repositories != null && selector != null) {
-            for (ArtifactRepository repository : repositories) {
-                repository.setProxy(getProxy(selector, repository));
-            }
-        }
-    }
-
-    private org.apache.maven.repository.Proxy getProxy(ProxySelector selector, ArtifactRepository repository) {
-        if (selector != null) {
-            RemoteRepository repo = RepositoryUtils.toRepo(repository);
-            org.eclipse.aether.repository.Proxy proxy = selector.getProxy(repo);
-            if (proxy != null) {
-                org.apache.maven.repository.Proxy p = new org.apache.maven.repository.Proxy();
-                p.setHost(proxy.getHost());
-                p.setProtocol(proxy.getType());
-                p.setPort(proxy.getPort());
-                if (proxy.getAuthentication() != null) {
-                    repo = new RemoteRepository.Builder(repo).setProxy(proxy).build();
-                    AuthenticationContext authCtx = AuthenticationContext.forProxy(null, repo);
-                    p.setUserName(authCtx.get(AuthenticationContext.USERNAME));
-                    p.setPassword(authCtx.get(AuthenticationContext.PASSWORD));
-                    p.setNtlmDomain(authCtx.get(AuthenticationContext.NTLM_DOMAIN));
-                    p.setNtlmHost(authCtx.get(AuthenticationContext.NTLM_WORKSTATION));
-                    authCtx.close();
-                }
-                return p;
-            }
-        }
-        return null;
-    }
-
-    private void injectAuthentication(AuthenticationSelector selector, List<ArtifactRepository> repositories) {
-        if (repositories != null && selector != null) {
-            for (ArtifactRepository repository : repositories) {
-                repository.setAuthentication(getAuthentication(selector, repository));
-            }
-        }
-    }
-
-    private Authentication getAuthentication(AuthenticationSelector selector, ArtifactRepository repository) {
-        if (selector != null) {
-            RemoteRepository repo = RepositoryUtils.toRepo(repository);
-            org.eclipse.aether.repository.Authentication auth = selector.getAuthentication(repo);
-            if (auth != null) {
-                repo = new RemoteRepository.Builder(repo)
-                        .setAuthentication(auth)
-                        .build();
-                AuthenticationContext authCtx = AuthenticationContext.forRepository(null, repo);
-                Authentication result = new Authentication(
-                        authCtx.get(AuthenticationContext.USERNAME), authCtx.get(AuthenticationContext.PASSWORD));
-                result.setPrivateKey(authCtx.get(AuthenticationContext.PRIVATE_KEY_PATH));
-                result.setPassphrase(authCtx.get(AuthenticationContext.PRIVATE_KEY_PASSPHRASE));
-                authCtx.close();
-                return result;
-            }
-        }
-        return null;
     }
 }
