+++ maven-core/src/test/java/org/apache/maven/internal/aether/DefaultRepositorySystemSessionFactoryTest.java
@@ -22,6 +22,7 @@ import javax.inject.Inject;
 
 import java.io.File;
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.Properties;
@@ -89,7 +90,8 @@ public class DefaultRepositorySystemSessionFactoryTest {
                 eventSpyDispatcher,
                 information,
                 defaultTypeRegistry,
-                versionScheme);
+                versionScheme,
+                Collections.emptyMap());
 
         MavenExecutionRequest request = new DefaultMavenExecutionRequest();
         request.setLocalRepository(getLocalRepository());
@@ -113,7 +115,8 @@ public class DefaultRepositorySystemSessionFactoryTest {
                 eventSpyDispatcher,
                 information,
                 defaultTypeRegistry,
-                versionScheme);
+                versionScheme,
+                Collections.emptyMap());
 
         MavenExecutionRequest request = new DefaultMavenExecutionRequest();
         request.setLocalRepository(getLocalRepository());
@@ -149,7 +152,8 @@ public class DefaultRepositorySystemSessionFactoryTest {
                 eventSpyDispatcher,
                 information,
                 defaultTypeRegistry,
-                versionScheme);
+                versionScheme,
+                Collections.emptyMap());
 
         PlexusConfiguration plexusConfiguration = (PlexusConfiguration) systemSessionFactory
                 .newRepositorySession(request)
@@ -193,7 +197,8 @@ public class DefaultRepositorySystemSessionFactoryTest {
                 eventSpyDispatcher,
                 information,
                 defaultTypeRegistry,
-                versionScheme);
+                versionScheme,
+                Collections.emptyMap());
 
         Map<String, String> headers = (Map<String, String>) systemSessionFactory
                 .newRepositorySession(request)
@@ -231,7 +236,8 @@ public class DefaultRepositorySystemSessionFactoryTest {
                 eventSpyDispatcher,
                 information,
                 defaultTypeRegistry,
-                versionScheme);
+                versionScheme,
+                Collections.emptyMap());
 
         int connectionTimeout = (Integer) systemSessionFactory
                 .newRepositorySession(request)
@@ -273,7 +279,8 @@ public class DefaultRepositorySystemSessionFactoryTest {
                 eventSpyDispatcher,
                 information,
                 defaultTypeRegistry,
-                versionScheme);
+                versionScheme,
+                Collections.emptyMap());
 
         int connectionTimeout = (Integer) systemSessionFactory
                 .newRepositorySession(request)
@@ -309,7 +316,8 @@ public class DefaultRepositorySystemSessionFactoryTest {
                 eventSpyDispatcher,
                 information,
                 defaultTypeRegistry,
-                versionScheme);
+                versionScheme,
+                Collections.emptyMap());
 
         int requestTimeout = (Integer) systemSessionFactory
                 .newRepositorySession(request)
@@ -351,7 +359,8 @@ public class DefaultRepositorySystemSessionFactoryTest {
                 eventSpyDispatcher,
                 information,
                 defaultTypeRegistry,
-                versionScheme);
+                versionScheme,
+                Collections.emptyMap());
 
         int requestTimeout = (Integer) systemSessionFactory
                 .newRepositorySession(request)
@@ -370,7 +379,8 @@ public class DefaultRepositorySystemSessionFactoryTest {
                 eventSpyDispatcher,
                 information,
                 defaultTypeRegistry,
-                versionScheme);
+                versionScheme,
+                Collections.emptyMap());
 
         MavenExecutionRequest request = new DefaultMavenExecutionRequest();
         request.setLocalRepository(getLocalRepository());
@@ -417,7 +427,8 @@ public class DefaultRepositorySystemSessionFactoryTest {
                 eventSpyDispatcher,
                 information,
                 defaultTypeRegistry,
-                versionScheme);
+                versionScheme,
+                Collections.emptyMap());
 
         MavenExecutionRequest request = new DefaultMavenExecutionRequest();
         request.setLocalRepository(getLocalRepository());
